# apps/device-management/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Smart Home API",
        "description": "REST / Kafka gateway for smart-home platform",
        "version": "2.0"
    },
    "paths": {
...
        "/api/v2.0/sensors/": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "List Devices",
                "operationId": "list_devices_api_v2_0_sensors__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/DeviceResponse"
                                    },
                                    "type": "array",
                                    "title": "Response List Devices Api V2 0 Sensors  Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Create Device",
                "operationId": "create_device_api_v2_0_sensors__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeviceBase"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2.0/sensors/{device_id}": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get Device",
                "operationId": "get_device_api_v2_0_sensors__device_id__get",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Device Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Update Device",
                "operationId": "update_device_api_v2_0_sensors__device_id__put",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Device Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeviceBase"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Delete Device",
                "operationId": "delete_device_api_v2_0_sensors__device_id__delete",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Device Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2.0/sensors/activate": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Activate Device",
                "operationId": "activate_device_api_v2_0_sensors_activate_post",
                "parameters": [
                    {
                        "name": "activation_code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Activation Code"
                        }
                    },
                    {
                        "name": "home_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Home Id"
                        }
                    },
                    {
                        "name": "room_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Room Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DeviceBase": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "type": {
                        "$ref": "#/components/schemas/DeviceType"
                    },
                    "model": {
                        "type": "string",
                        "title": "Model"
                    },
                    "firmware_version": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Firmware Version"
                    },
                    "status": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Status"
                    },
                    "room_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid4"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Room Id"
                    },
                    "home_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid4"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Home Id"
                    },
                    "activation_code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Activation Code"
                    },
                    "is_activated": {
                        "type": "boolean",
                        "title": "Is Activated",
                        "default": false
                    },
                    "activated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Activated At"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "type",
                    "model"
                ],
                "title": "DeviceBase"
            },
            "DeviceResponse": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "type": {
                        "$ref": "#/components/schemas/DeviceType"
                    },
                    "model": {
                        "type": "string",
                        "title": "Model"
                    },
                    "firmware_version": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Firmware Version"
                    },
                    "status": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Status"
                    },
                    "room_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid4"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Room Id"
                    },
                    "home_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid4"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Home Id"
                    },
                    "activation_code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Activation Code"
                    },
                    "is_activated": {
                        "type": "boolean",
                        "title": "Is Activated",
                        "default": false
                    },
                    "activated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Activated At"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "type",
                    "model",
                    "id"
                ],
                "title": "DeviceResponse"
            },
            "DeviceType": {
                "type": "string",
                "enum": [
                    "SENSOR",
                    "CAMERA",
                    "SWITCH"
                ],
                "title": "DeviceType"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}